Removed Duplicate CarList.test
